Добавляет Модуляцию JS
