+++ js/util.js
@@ -1,0 +1,26 @@
+// Функция для генерации случайного числа в заданном диапазоне
+function getRandomInt(min, max) {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+// Функция для выбора случайного элемента из массива
+const getRandomElement = (array) => array[getRandomInt(0, array.length - 1)];
+
+function getId (min, max) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInt(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      window.console.error(`Перебраны все числа из диапазона от ${ min } до ${ max}`);
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInt(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+export {getRandomElement, getId, getRandomInt};
